Refactor test utils package name
